Move flowlayout to a separate file
