+++ lib/widget/paragraph.cpp
@@ -26,203 +26,6 @@
 #include "form.h"
 #include "tip.h"
 
-class FlowLayoutElementDescriptor
-{
-public:
-    virtual ~FlowLayoutElementDescriptor() = default;
-    virtual unsigned int getWidth(size_t position, size_t length) const = 0;
-    virtual size_t size() const = 0;
-    virtual bool isWhitespace(size_t position) const = 0;
-    virtual bool isLineBreak(size_t position) const = 0;
-
-    virtual bool isWord(size_t position) const
-    {
-        return position < size() && !isWhitespace(position) && !isLineBreak(position);
-    }
-};
-
-/**
- * FlowLayout implements a word wrapping algorithm.
- *
- * It doesn't specify how the text should be structured, it only requires that the elements of the text are described
- * by FlowLayoutElementDescriptor.
- *
- * Each element can contain 0 or more words, and a single word can be part of two sequential elements.
- *
- * Example:
- * - Element 1 = "first sec"
- * - Element 2 = "ond third"
- *
- * In the example above, the resulting text is "first second third", so if possible the algorithm will keep
- * the second word completely in the same line.
- **/
-struct FlowLayout
-{
-private:
-    bool shouldMergeFragment()
-    {
-        return !currentLine.empty() && !partialWord.empty() && currentLine.back().elementId == partialWord.front().elementId;
-    }
-
-    void endWord()
-    {
-        if (shouldMergeFragment())
-        {
-            auto partialWordFront = &partialWord.front();
-            auto currentLineBack = currentLine.back();
-            currentLine.pop_back();
-            partialWordFront->length = partialWordFront->length + partialWordFront->begin - currentLineBack.begin;
-            partialWordFront->width = partialWordFront->offset + partialWordFront->width - currentLineBack.offset;
-            partialWordFront->begin = currentLineBack.begin;
-            partialWordFront->offset = currentLineBack.offset;
-        }
-
-        for (auto fragment: partialWord)
-        {
-            currentLine.push_back(fragment);
-        }
-
-        nextOffset += partialWordWidth;
-        partialWordWidth = 0;
-        partialWord.clear();
-    }
-
-    void pushFragment(FlowLayoutElementDescriptor const &elementDescriptor, size_t begin, size_t end)
-    {
-        auto width = elementDescriptor.getWidth(begin, end - begin);
-        partialWord.push_back({currentElementId, begin, end - begin, width, nextOffset + partialWordWidth});
-        partialWordWidth += width;
-    }
-
-    void placeLine(FlowLayoutElementDescriptor const &elementDescriptor, size_t begin, size_t end)
-    {
-        auto current = begin;
-
-        while (current < end)
-        {
-            long long fragmentFits;
-
-            if (nextOffset + partialWordWidth + elementDescriptor.getWidth(current, end - current) > maxWidth)
-            // fragment doesn't fit completely in the current line
-            {
-                fragmentFits = current - 1;
-                size_t fragmentDoesntFit = end;
-                while (fragmentDoesntFit - fragmentFits > 1)
-                {
-                    auto middle = (fragmentFits + fragmentDoesntFit) / 2;
-                    if (nextOffset + partialWordWidth + elementDescriptor.getWidth(current, middle - current) > maxWidth)
-                    {
-                        fragmentDoesntFit = middle;
-                    } else {
-                        fragmentFits = middle;
-                    }
-                }
-            }
-            else
-            {
-                fragmentFits = end;
-            }
-
-            auto whitespacePosition = fragmentFits + 1;
-            while ((whitespacePosition >= elementDescriptor.size() || !elementDescriptor.isWhitespace(whitespacePosition - 1)) && whitespacePosition > current)
-            {
-                whitespacePosition--;
-            }
-
-            if (whitespacePosition > current)
-            // the fragment ending with a whitespace fits within the line
-            {
-                pushFragment(elementDescriptor, current, whitespacePosition - 1);
-                endWord();
-                nextOffset += elementDescriptor.getWidth(whitespacePosition - 1, 1);
-                current = whitespacePosition;
-            }
-            else if (fragmentFits == end)
-            // the fragment is a single word, and fits in the line,
-            // but it doesn't end in whitespace and the next element might be part of this word
-            {
-                pushFragment(elementDescriptor, current, fragmentFits);
-                current = fragmentFits;
-            }
-            else if (nextOffset > 0)
-            // word doesn't fit in the current line
-            {
-                 breakLine();
-            }
-            else
-            // word doesn't fit in an empty line
-            {
-                auto fragmentEnd = std::max((size_t)fragmentFits, current + 1);
-                pushFragment(elementDescriptor, current, fragmentEnd);
-                current = fragmentEnd;
-                breakLine();
-            }
-        }
-    }
-
-public:
-    FlowLayout(unsigned int maxWidth): maxWidth(maxWidth)
-    {
-
-    }
-
-    void append(FlowLayoutElementDescriptor const &elementDescriptor)
-    {
-        size_t position = 0;
-        while (position < elementDescriptor.size())
-        {
-            auto lineEnd = position;
-
-            while (lineEnd < elementDescriptor.size() && !elementDescriptor.isLineBreak(lineEnd))
-            {
-                lineEnd++;
-            }
-
-            placeLine(elementDescriptor, position, lineEnd);
-
-            if (lineEnd < elementDescriptor.size())
-            {
-                breakLine();
-            }
-
-            position = lineEnd + 1;
-        }
-
-        currentElementId++;
-    }
-
-    void end()
-    {
-        endWord();
-
-        if (!currentLine.empty()) {
-            breakLine();
-        }
-    }
-
-    void breakLine()
-    {
-        endWord();
-        lines.push_back(currentLine);
-        currentLine.clear();
-        nextOffset = 0;
-    }
-
-    std::vector<std::vector<FlowLayoutFragment>> getLines()
-    {
-        return lines;
-    }
-
-private:
-    std::vector<FlowLayoutFragment> currentLine;
-    std::vector<std::vector<FlowLayoutFragment>> lines;
-    unsigned int maxWidth;
-    unsigned int currentElementId = 0;
-    unsigned int nextOffset = 0;
-    std::vector<FlowLayoutFragment> partialWord;
-    unsigned int partialWordWidth = 0;
-};
-
 /**
  * Used to render a fragment of text in the paragraph.
  *
@@ -255,20 +58,6 @@ public:
 private:
 	WzCachedText cachedText;
 	ParagraphTextStyle textStyle;
-};
-
-/**
- * Provide information about an element appended to the paragraph.
- **/
-struct ParagraphElement
-{
-	virtual ~ParagraphElement() = default;
-
-	virtual void appendTo(FlowLayout &layout) = 0;
-	virtual WIDGET *createFragmentWidget(Paragraph &paragraph, FlowLayoutFragment const &fragment) = 0;
-	virtual void destroyFragments(Paragraph &paragraph) = 0;
-	virtual bool isLayoutDirty() const = 0;
-	virtual int32_t getAboveBase() const = 0;
 };
 
 class FlowLayoutStringDescriptor : public FlowLayoutElementDescriptor
